pyproject.toml
[project]
name = "ai-auto-consultant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-groq>=0.3.7",
    "langgraph>=0.6.4",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]
